my_ui_agent/requirements.txt
# Flask веб-фреймворк
Flask==3.0.0
werkzeug==3.0.1

# Google Cloud Vision API
google-cloud-vision==3.4.4

# AI Vision модели
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.20.0
anthropic>=0.25.0

# Обработка изображений
Pillow==10.0.0
numpy==1.24.3
opencv-python>=4.8.0

# Дополнительные зависимости
python-dotenv>=1.0.0
requests>=2.31.0
asyncio>=3.4.3
aiohttp>=3.8.0

# Google OAuth и аутентификация
google-auth-oauthlib>=1.2.0
google-auth>=2.0.0

# Jupyter для тестирования
jupyter>=1.0.0
ipykernel>=6.0.0

# Логирование и мониторинг
colorlog>=6.0.0
tqdm>=4.65.0
